adapted to moved DI scopes
