+++ bundles/org.palladiosimulator.indirections.scheduler/src/org/palladiosimulator/indirections/scheduler/util/DataChannelResourceRegistry.java
@@ -12,10 +12,10 @@ import org.palladiosimulator.indirections.scheduler.IDataChannelResource;
 import org.palladiosimulator.indirections.scheduler.IDataChannelResourceFactory;
 import org.palladiosimulator.pcm.core.composition.AssemblyContext;
 import org.palladiosimulator.simulizar.di.component.interfaces.SimulatedThreadComponent;
+import org.palladiosimulator.simulizar.di.scopes.RuntimeExtensionScope;
 import org.palladiosimulator.simulizar.interpreter.InterpreterDefaultContext;
 import org.palladiosimulator.simulizar.interpreter.InterpreterDefaultContext.MainContext;
 import org.palladiosimulator.simulizar.runtimestate.RuntimeStateEntityManager;
-import org.palladiosimulator.simulizar.scopes.RuntimeExtensionScope;
 
 import de.uka.ipd.sdq.simucomframework.core.model.SimuComModel;
